Changed Maze to MazeDrawer Class, Created New Maze Class
MazeDrawer Class
-was the old maze class and removed addcell method
-changed the constructor parameters
-Allows a maze to be produced in game by pasing in a maze object in the
constructor
Maze Class
-created add cell method
-added method for getting List<MazeCell> as an object
