+++ VR Maze/Assets/Scripts/Maze.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: c7f8a017584a0314d880d9c475b31473
+timeCreated: 1488431319
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
